q-n-a/src/app/modules/authentication/authentication.service.ts
import { Observable } from 'rxjs';
import { Injectable } from '@angular/core';
import { HttpClient } from '@angular/common/http';
import { MatSnackBar } from '@angular/material/snack-bar';

import * as jwt_decode from 'jwt-decode';

export const TOKEN_NAME:string = 'jwt_token';

@Injectable()
export class AuthenticationService{

    springEndpoint: string;
    token: string;

    constructor(private httpClient: HttpClient,private snackBar : MatSnackBar){
        this.springEndpoint = "http://localhost:8089/user";
    }

    registerUser(newUser){
        const url = this.springEndpoint+"/register";
        console.log("new user in service",newUser);
        return this.httpClient.post(url,newUser,{responseType: 'text'});
    }

    loginUser(newUser){
        const url = this.springEndpoint+"/login";
        return this.httpClient.post(url,newUser);
    }

    setToken(token:string){
        return localStorage.setItem(TOKEN_NAME,token);
    }

    getToken(){
        return localStorage.getItem(TOKEN_NAME);
    }

    deleteToken(){
        return localStorage.removeItem(TOKEN_NAME);
    }

    getTokenExpirationDate(token: string): Date {
        const decoded = jwt_decode(token);
        if(decoded.exp == undefined) return null;
        const date  = new Date(0);
        date.setUTCSeconds(decoded.exp);
        return date;
    }

    isTokenExpired(token?:string): boolean {
        if(!token) token = this.getToken();
        if(!token) return true;
        const date = this.getTokenExpirationDate(token);
        if(date === undefined || date === null ) return false;
        return !(date.valueOf() > new Date().valueOf());
    }

    handleError(err : any){
        console.log(err);
        //console.log("error log "+err.error);
        this.snackBar.open(err.message , 'close');
    }
}